python/requirements.txt
dnspython==2.3.0
numpy==1.24.3
pandas==2.0.2
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
torchtext==0.14.1
torch==1.13.1
numpy
jupyter
notebook
matplotlib